requirements.txt
langchain
langchain-google-genai
google-generativeai
streamlit
chromadb
pandas
numpy
geopandas
shapely
folium
streamlit-folium
pillow
requests
deepeval
pydantic
python-dotenv
geopy
pyproj
